Fix #208 — 🐛 Check if provisioning lockfile mod time exists before using it

+++ lib/Lufi.pm
@@ -83,7 +83,7 @@ sub startup {
             my ($dev,$ino,$mode,$nlink,$uid,$gid,$rdev,$size,$atime,$mtime,$ctime,$blksize,$blocks) = stat($lockfile);
 
             # Remove the lockfile if more than 20 seconds old
-            if (time - $mtime > 20) {
+            if ($mtime && time - $mtime > 20) {
                 unlink $lockfile if -e $lockfile; # if -e just to be sure the file hasn’t been removed while checking it
             } else {
                 return;
